Fix system proxy bugs
