+++ shadowsocks-csharp/Proxy/SystemProxy/INTERNET_PER_CONN_OptionEnum.cs
@@ -1,0 +1,19 @@
+﻿namespace Shadowsocks.Proxy.SystemProxy
+{
+    /// <summary>
+    /// Constants used in INTERNET_PER_CONN_OPTION_OptionUnion struct.
+    /// </summary>
+    public enum INTERNET_PER_CONN_OptionEnum
+    {
+        INTERNET_PER_CONN_FLAGS = 1,
+        INTERNET_PER_CONN_PROXY_SERVER = 2,
+        INTERNET_PER_CONN_PROXY_BYPASS = 3,
+        INTERNET_PER_CONN_AUTOCONFIG_URL = 4,
+        INTERNET_PER_CONN_AUTODISCOVERY_FLAGS = 5,
+        INTERNET_PER_CONN_AUTOCONFIG_SECONDARY_URL = 6,
+        INTERNET_PER_CONN_AUTOCONFIG_RELOAD_DELAY_MINS = 7,
+        INTERNET_PER_CONN_AUTOCONFIG_LAST_DETECT_TIME = 8,
+        INTERNET_PER_CONN_AUTOCONFIG_LAST_DETECT_URL = 9,
+        INTERNET_PER_CONN_FLAGS_UI = 10
+    }
+}
